+++ packages/prisma-fabbrica/src/helpers/index.ts
@@ -1,7 +1,5 @@
 export * from "./valueResolver";
 export * from "./stringConverter";
-export * from "./astShorthand";
 export * from "./sequence";
 export * from "./selectors";
-export * from "./createJSONLiteral";
 export * from "./list";
