+++ packages/prisma-fabbrica/src/templates/ast-tools/astShorthand.ts
@@ -1,5 +1,5 @@
 import ts from "typescript";
-import { pascalize } from "./stringConverter";
+import { pascalize } from "../../helpers";
 
 type StripCreate<T extends string> = T extends `create${infer S}` ? Uncapitalize<S> : T;
